+++ src/app/[locale]/viewer/registry/_components/RegistryListView.tsx
@@ -4,7 +4,7 @@ import { useSearchParams } from 'next/navigation';
 import AssetNotFound from 'components/basics/AssetNotFound';
 import { encodeBase64 } from 'lib/util/Base64Util';
 import ListHeader from 'components/basics/ListHeader';
-import { performSearchAasFromAllRepositories } from 'lib/services/aas-repository-service/aasRepositorySearchActions';
+import { searchAasInAllRepositories } from 'lib/services/aas-repository-service/aasRepositorySearchActions';
 import { useTranslations } from 'next-intl';
 import { LocalizedError } from 'lib/util/LocalizedError';
 import { AasListEntry } from 'lib/types/AasListEntry';
@@ -29,7 +29,7 @@ export const RegistryListView = () => {
             throw new LocalizedError('pages.registryList.errors.noAasId');
         }
 
-        const response = await performSearchAasFromAllRepositories(encodeBase64(aasId));
+        const response = await searchAasInAllRepositories(encodeBase64(aasId));
 
         if (!response.isSuccess) {
             throw new LocalizedError('pages.registryList.errors.searchFailed');
