+++ src/lib/services/discovery-service/DiscoveryService.ts
@@ -4,13 +4,13 @@ import {
     wrapErrorCode,
     wrapSuccess,
 } from 'lib/util/apiResponseWrapper/apiResponseWrapper';
-import { InfrastructureConnection } from 'lib/services/infrastructure-search-service/InfrastructureSearchService';
 import { ApiResultStatus } from 'lib/util/apiResponseWrapper/apiResultStatus';
 import logger, { logResponseDebug } from 'lib/util/Logger';
 import { IDiscoveryServiceApi } from 'lib/api/discovery-service-api/discoveryServiceApiInterface';
 import { mnestixFetch } from 'lib/api/infrastructure';
 import { DiscoveryServiceApi } from 'lib/api/discovery-service-api/discoveryServiceApi';
 import { getInfrastructuresIncludingDefault } from 'lib/services/database/connectionServerActions';
+import { InfrastructureConnection } from 'lib/services/database/MappedTypes';
 
 export type DiscoverySearchResult = {
     aasId: string;
