+++ src/lib/services/submodel-registry-service/SubmodelRegistryService.ts
@@ -5,9 +5,9 @@ import { ApiResponseWrapper, wrapErrorCode, wrapSuccess } from 'lib/util/apiResp
 import { SubmodelDescriptor } from 'lib/types/registryServiceTypes';
 import { ApiResultStatus } from 'lib/util/apiResponseWrapper/apiResultStatus';
 import { Submodel } from 'lib/api/aas/models';
-import { InfrastructureConnection } from 'lib/services/infrastructure-search-service/InfrastructureSearchService';
 import { fetchFromMultipleEndpoints } from 'lib/services/shared/parallelFetch';
 import { ISubmodelRegistryServiceApi } from 'lib/api/submodel-registry-service/submodelRegistryServiceApiInterface';
+import { InfrastructureConnection } from 'lib/services/database/MappedTypes';
 
 export class SubmodelRegistryService {
     private constructor(
